Download, translate and read OSM extracts

Download, translate and read OSM extracts

oe_get(
  place,
  layer = "lines",
  provider = "geofabrik",
  ...,
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  download_directory = oe_download_directory(),
  force_download = FALSE,
  max_file_size = 5e+08,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  download_only = FALSE,
  oe_verbose = FALSE,
  oe_quiet = FALSE
)

Arguments

place
Description of the geographical area that should be download through the chosen provider. Can be either a length-1 character vector, a length-1 sfc_POINT object or a numeric vector with length 2. See details and examples.

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations.

provider
Which provider should be used to download the data? Available providers can be found with the following command: osmextractr:::oe_available_providers()

...
Arguments that should be passed to sf::st_read()

match_by
Which column of the provider data should be used for matching the input place with the provider's data? The default is "name". Check details and examples to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numeric value greater or equal than 0. What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

interactive_ask
Boolean. If TRUE the function creates and interactive menu in case the best match is further than max_string_dist. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

download_directory
Where to download the file containing the OSM data? By default this is tempdir(), which changes each time you restart R. You can set a persistent download_directory by adding the following to your .Renviron file (e.g. with usethis::edit_r_environ()): OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data.

force_download
Should the file be updated if it has already been downloaded? FALSE by default.

max_file_size
The maximum file size to download without asking in interactive mode. Default: 5e+8, half a gigabyte.

vectortranslate_options
Options to pass to the sf::gdal_utils() argument options. Set by default. Check details and examples.

osmconf_ini
The configuration file specifying which columns should be in the resulting data frame. See documentation at gdal.org for details.

extra_attributes
Which addition columns, corresponding to OSM keys, should be in the resulting dataset? NULL by default.

force_vectortranslate
Force the original .pbf file to be translated into a .gpkg file, even if a .gpkg associated with the provider zone already exists.

download_only
Boolean. If TRUE then the function only returns the path where the matched file is stored, instead of reading it. FALSE by default.

oe_verbose
Boolean. If TRUE the function prints informative messages.

oe_quiet
Should files be downloaded without a progress bar? FALSE by default.

Value

An sf object related to the input place.

Details

This function is a wrapper around ...

Examples

iow = oe_get("Isle of Wight", provider = "test", oe_verbose = TRUE)

#> The input place was matched with: Isle of Wight

#> File downloaded!

#> Start with the vectortranslate operations on the input file!

#> Finished the vectortranslate operations on the input file!

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNLZrOn/test_geofabrik_isle-of-wight-latest.gpkg' using driver `GPKG'
#> Simple feature collection with 44365 features and 9 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -5.401978 ymin: 43.35489 xmax: -0.175775 ymax: 50.89599
#> CRS:            4326

class(iow)

#> [1] "sf"         "data.frame"

summary(sf::st_geometry_type(iow))

#>           GEOMETRY              POINT         LINESTRING            POLYGON 
#>                  0                  0              44365                  0 
#>         MULTIPOINT    MULTILINESTRING       MULTIPOLYGON GEOMETRYCOLLECTION 
#>                  0                  0                  0                  0 
#>     CIRCULARSTRING      COMPOUNDCURVE       CURVEPOLYGON         MULTICURVE 
#>                  0                  0                  0                  0 
#>       MULTISURFACE              CURVE            SURFACE  POLYHEDRALSURFACE 
#>                  0                  0                  0                  0 
#>                TIN           TRIANGLE 
#>                  0                  0 

iow_points = oe_get("Isle of Wight", provider = "test", layer = "points")

#> Reading layer `points' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNLZrOn/test_geofabrik_isle-of-wight-latest.gpkg' using driver `GPKG'
#> Simple feature collection with 58945 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -3.805609 ymin: 43.45894 xmax: -1.046749 ymax: 50.89437
#> CRS:            4326

summary(sf::st_geometry_type(iow_points))

#>           GEOMETRY              POINT         LINESTRING            POLYGON 
#>                  0              58945                  0                  0 
#>         MULTIPOINT    MULTILINESTRING       MULTIPOLYGON GEOMETRYCOLLECTION 
#>                  0                  0                  0                  0 
#>     CIRCULARSTRING      COMPOUNDCURVE       CURVEPOLYGON         MULTICURVE 
#>                  0                  0                  0                  0 
#>       MULTISURFACE              CURVE            SURFACE  POLYHEDRALSURFACE 
#>                  0                  0                  0                  0 
#>                TIN           TRIANGLE 
#>                  0                  0 

# Get the .pbf file path
oe_get("Isle of Wight", download_only = TRUE)

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpNLZrOn/geofabrik_isle-of-wight-latest.gpkg"

if (FALSE) {
# alternative providers
baku = oe_get(place = "Baku", provider = "bbbike", oe_verbose = TRUE)
}