Match input place with a geographical zone

This function is used to match the input place with the URL of the corresponding pbf file (and its file size, if present).

oe_match(place, ...)

# S3 method for default
oe_match(place, ...)

# S3 method for sfc_POINT
oe_match(place, provider = "geofabrik", quiet = TRUE, ...)

# S3 method for numeric
oe_match(place, provider = "geofabrik", quiet = TRUE, ...)

# S3 method for character
oe_match(
  place,
  provider = "geofabrik",
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  quiet = TRUE,
  ...
)

Arguments

place
Description of the geographical area that should be download through the chosen provider. Can be either a length-1 character vector, a length-1 sfc_POINT object or a numeric vector with length 2. See details and examples.

...
arguments passed to other methods

provider
Which provider should be used to download the data? Available providers can be found with the following command: osmextractr:::oe_available_providers()

quiet
Boolean. If FALSE the function prints informative messages.

match_by
Which column of the provider data should be used for matching the input place with the provider's data? The default is "name". Check details and examples to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numeric value greater or equal than 0. What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

interactive_ask
Boolean. If TRUE the function creates and interactive menu in case the best match is further than max_string_dist. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

Value

A list with two elements, named url and file_size. The first element is the URL of the file associated with the input place, while the second element is the size of the file (which may be NULL or NA)

See also

Examples

# The simplest example:
oe_match("Italy")

#> $url
#> [1] "https://download.geofabrik.de/europe/italy-latest.osm.pbf"
#> 
#> $file_size
#> [1] 1544340778
#> 

# Don't run example, see issue #49
# Match the input zone using an sfc_POINT object:
milan_duomo = sf::st_sfc(sf::st_point(c(1514924, 5034552)), crs = 3003)
oe_match(milan_duomo)

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#> 

# Match the input zone using a numeric vector of coordinates
# (in which case crs = 4326 is assumed)
oe_match(c(9.1916, 45.4650)) # Milan, Duomo using CRS = 4326

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#> 

# Perform the matching operations using other columns instead of "name".
oe_match("RU", match_by = "iso3166_1_alpha2")

#> $url
#> [1] "https://download.geofabrik.de/russia-latest.osm.pbf"
#> 
#> $file_size
#> [1] 2820253009
#> 

# Increase the max_string_dist parameter and help the function:
oe_match("Isle Wight", max_string_dist = 3)

#> $url
#> [1] "https://download.geofabrik.de/europe/great-britain/england/isle-of-wight-latest.osm.pbf"
#> 
#> $file_size
#> [1] 6877468
#> 

# but be aware that it can be dangerous:
oe_match("London", max_string_dist = 3, quiet = FALSE)

#> The input place was matched with: Jordan

#> $url
#> [1] "https://download.geofabrik.de/asia/jordan-latest.osm.pbf"
#> 
#> $file_size
#> [1] 27400228
#> 

# Check interactive_ask:
if (interactive()) {
    oe_match("London", interactive_ask = TRUE)
}

# Change the provider:
oe_match("Leeds", provider = "bbbike")

#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Leeds/Leeds.osm.pbf"
#> 
#> $file_size
#> [1] 19376705
#>