Download, translate and read OSM extracts

Download, translate and read OSM extracts

osmext_get(
  place,
  layer = "lines",
  provider = "geofabrik",
  ...,
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  download_directory = osmext_download_directory(),
  force_download = FALSE,
  max_file_size = 5e+08,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  osmext_verbose = FALSE
)

Arguments

place
Description of the geographical area that should be download through the chosen provider. Can be either a length-1 character vector, a length-1 sfc_POINT object or a numeric vector with length 2. See details and examples.

layer
TODO

provider
Which provider should be used to download the data? For the moment we support only "geofabrik".

...
Arguments that should be passed to sf::st_read()

match_by
Which column of the provider data should be used for matching the input place with the provider's data? The default is "name". Check details and examples to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numeric value greater or equal than 0. What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

interactive_ask
Boolean. If TRUE the function creates and interactive menu in case the best match is further than max_string_dist. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

download_directory
TODO

force_download
TODO

max_file_size
TODO

vectortranslate_options
TODO

osmconf_ini
TODO

extra_attributes
TODO

force_vectortranslate
TODO

osmext_verbose
Boolean. If TRUE the function prints informative messages.

Value

An sf object related to the input place.

Details

This function is a wrapper around ...

Examples

osmext_get("Isle of Wight", provider = "test")

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp1wJWmQ/geofabrik_geofabrik_isle-of-wight-latest.gpkg' using driver `GPKG'
#> Simple feature collection with 44365 features and 9 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -5.401978 ymin: 43.35489 xmax: -0.175775 ymax: 50.89599
#> CRS:            4326